RRC: Fix pointer-sign issue on ENUMERATED types
The VAL_PTR returned on an ENUMERATED types is always unsigned.
(This is not necessarily true in general, as NamedNumbers in ENUMERATED
types have a SignedNumber value and can be negative, but there are
no examples in the asn in the repository. Still, perhaps asn2wrs.py
should be updated for this situation.)

Fix the pointer-sign issues related to that. (None of these can
be larger than INT32_MAX, so a signed type is sometimes used to
allow -1 to represent unknown.)

Ping #20175