uat: Fix types involved with uat_esc and uat_unesc
uat_esc and uat_unesc escape and unescape strings, converting
between an array of bytes typically mostly printable ASCII but
possibly containing other character, including internal null bytes,
and a null-terminated ASCII string using C-style escapes.

They are used, for example, with the LSTRING macros that can be
used for things like encryption keys that are treated as bytes but
are often mostly strings. The unescaped array of bytes type should
presumably *not* be a char* since it's not a null terminated string
and will cause problems if used with certain string standard library
functions.

Fix the types on those functions. This prevents pointer-sign warnings
with the UAT associated with the SMTP dissector.

Also fixup 81ec7d15f0f34a8d8599c76d15f7877db856d033 some for this;
since it uses this type (which can then contain an internal null), it
should not use strlen and it should use lua_pushlstring with the length.

uat_undquote and uat_unbinstring should probably return uint8_t* as
well, but that would require more code changes, such as in uat_load.l

Ping #20175