Add extern "C" for C++ compatibility
epan/reassemble.h and epan/ptvcursor.h didn't have extern "C" for public
function which resulted in attempt to link C++ mangled names when used
from C++ source file.

For reassemble.h it was possible to wrap #include in extern "C" but
ptvcursor.h proved more difficult - packet.h includes glib header that,
when in C++ and on Win32 used template<> which is a failure when in extern
"C" block.