ASN.1 BER dissector: dissect_ber_integer pointer-sign warnings
dissect_ber_integer always returns a uint32_t for its VAL_PTR
variable, despite the default for an unconstrained IntegerType
being FT_INT32.

Note that dissect_ber_integer always sets the VAL_PTR to 0
upon entering; setting the variable to -1 and testing it later
does not work the way it does for CHOICEs and dissect_ber_choice,
so that idiom can be removed.

Some of these are actually ENUMERATED, which are always FT_UINT32,
because ENUMERATED is encoded just as an INTEGER in the BER.

Some of these are actually FT_UINT32 for other reasons, like the
type manually being overridden.

Some of these are FT_INT32 because that's the default and there's
no constraint so a negative could somehow be encoded, but *probably*
are unsigned in practice because they have a NamedNumberList that
has a small selection of unsigned values. (They *should* probably be
either ENUMERATED or have a ValueRange constraint, but that's an issue
in the original ASN.1)

This is the last set of pointer-sign warnings for #20175