pkcs12: Put a maximum limit on hash iterations willing to perform
iterationCount is defined in the ASN.1 for pkcs12 (RFC 8081) as
an unbounded Integer (either (1..MAX) in one place, or with no
value constraint in another, though the latter is silly as negative
numbers make no sense.) We should have a sanity check on it to avoid
endless hashing that consumes an incredible amout of time (especially
as the iteration_count value is eventually cast to unsigned, so a -1
becomes whatever the maximum size of that unsigned variable is).

RFC 8081 Section 4.2 Iteration Count says:

   Choosing a reasonable value for the iteration count depends on
   environment and circumstances, and varies from application to
   application.  This document follows the recommendations made in FIPS
   Special Publication 800-132 [NISTSP132], which says

      The iteration count shall be  selected as large as possible, as
      long as the time required to generate the key using the entered
      password is acceptable for the users. [...] A minimum iteration
      count of 1,000 is recommended.  For especially critical keys, or
      for very powerful systems or systems where user-perceived
      performance is not critical, an iteration count of 10,000,000 may
      be appropriate.

While a user may want to decrypt "especially critical keys", OTOH
generally "user-perceived performance" is of a concern.

This puts a limit of 10,000,000 instead of allowing 2*32 as we've been
doing, though we could make it an unsigned preference at some point.

This also happens to fix a pointer-sign warning by changing the variable
to unsigned earlier (as opposed to casting it when calling
generate_key_or_iv).

Ping #20175