X11: Make sure we don't overflow when stepping to the next message
The X11 dissector retrieves a length for each message, but, depending
on the opcode, doesn't necessarily try to access frame data and add
dissection items for all of those bytes. (Maybe it should warn about
inconsistent lengths.) This means that it might not throw various
BoundsErrors before attempting to step to the next message. Make
sure we don't overflow the offset and break out of the loop if so.

OSS-Fuzz 461640135