Homeplug AV: Add to COL_INFO even with a non-NULL tree
The Homeplug AV dissector does a lot of checking for a NULL tree.
In one path, text is appended to COL_INFO only when there *is* a
NULL tree, but we can have a non-NULL tree and also need to fill
in the columns in a number of situations, notably if there are
custom columns.

Append the text in the other code path. Just get rid of the NULL
tree check, any savings are minimal due to the NULL tree checks
inside the API.

Fix #20893