jtckdint: Workaround C _Generic lvalue conversion
The controlling expression of a C11 _Generic undergoes lvalue
conversion (unlike, say, C++ templates):

https://www.open-std.org/jtc1/sc22/wg14/www/docs/n2396.htm#dr_481

Which means that the volatile qualifier is removed. This results,
with the current implementation, in undefined behavior when writing
to a volatile result through a non-volatile pointer, and a MSVC
C4090 warning.

https://learn.microsoft.com/en-us/cpp/error-messages/compiler-warnings/compiler-warning-level-1-c4090

It is not UB or a problem to cast to a volatile pointer to add the
qualifier (other than possible small performance implications), so
do that. A more complicated rewrite of the macros could eliminate
the need to do that.

C++ (as it uses templates), any C or C++ library using the C23 standard
header, and any use the gcc/clang built-ins are unaffected by this
change.