RADIUS: Increment include depth before testing value
Increment the include depth before testing if it has reached its
max value. (This is the only place it is incremented, so using its
current value to assign the current buffer and post-incrementing is
ok.)

As it is, this test is entirely ineffective. (It is unlikely that
an include depth of 10 will occur.)

Coverity CIDs 1518994, 1518996