MIDI: Move packet-sysex.c to packet-midi.c.
We really need a home for MIDI functionality shared/shareable between RTP-MIDI
and USBAudio MIDI.  At the same time, it seems a bit weird having just MIDI
SysEx stuff in its own standalone file, so let's use the SysEx code as the
starting point for the common MIDI code, moving it from packet-sysex.c to
packet-midi.c for that purpose.

Just rearranging files (and minimally updating the file header comments) for
now.  Further refactoring and adjustments to follow...

This is followup from discussion in !22141.