HTTP3: Fix quarter stream too big calculation
The shift is in the wrong direction. A QUIC stream ID is 62-bit, and
a Quarter Stream (stream ID divided by 4) is 60-bit. (Luckily the
way the test is written, the fault results in never printing the expert
info instead of resulting in always printing the expert info and
aborting.)

Coverity CID 1662012