observer: Fix saving files
The tlv_time_info struct does not need to contain the type and length
fields, which are already contained in the tlv_header struct. They
are never initialized, and their size is erroneously used to increase
the TLV size indicated in the TLV header that is written before the
extra type and length fields. This causes any file written by libwiretap
to fail to open, with a message like:

        The file "observer_bad.bfr" appears to be damaged or corrupt.
        (Observer: bad record (time information TLV length 12 != 8))

Cf. with the other tlv value structures, like tlv_network_load,
tlv_wireless_info, and tlv_capture_start_stop, none of which contain
the T or L from the TLV, just the V.

Fix writing Network Instruments/JDSU/Viavi Observer files

Coverity CID 1499434