Qt: Display all the MEGACO and MGCP Response Time Delay timestats
There are two types of rtd_tables; ones where num_rtds == 1 and
num_timestat > 1 (so that the number of duplicate/open/discarded
requests and responses are shared among all the types), and ones
where num_rtds > 1 and num_timestat == 1 for each rtd_timestat
(so that each type has its own separate accounting of duplicate/
open/discarded requests and responses.)

MEGACO and MGCP are the former; H.225 and RADIUS are the latter.
The former are not being displayed correctly in the Qt dialog,
although they are handled correctly by the CLI tap-rtd.c.
As a result, only the results for MEGACO Add commands are being
shown in the GUI.

(See for example the file FAX-Call-t38-CA-TDM-SIP-FB-1.pcap on
the Sample Captures page on the wiki.)

Display the other rows as well. Note that MEGACO and MGCP make a
different decision whether a total/all type is the first or last
timestat within the rtd. If they were the same, then perhaps we
could make the other types children of the total type. (That would
make somewhat more sense with how the duplicates are presented.)