MIDI: Adjust naming to reflect that SysEx is MIDI.
This commit adjusts some names and strings as followup to 1918de0629, helping
prepare for integrating common MIDI dissection code into packet-midi.c .

This commit includes some changes to filter strings for which backwards
compatibility is probably not nearly as valuable as future correctness.

Replace "SYSEX" with "SysEx" too, as that's how it is capitalized in the MIDI
spec.

Replace "command" with "message" in the places where its usage is generic; The
MIDI spec does contain a mixture of usages of "message" and "command" with no
apparent distinction, but has a very strong bias towards "message".  Let's
standardize on "message".  (The DigiTech SysEx messages might well contain
"commands", so those usages are left as-is.)