UMTS RLC: Make Coverity happy about initializing members
Comment out a rlc_seq struct member that isn't used. Initialize
a struct member before copying it to a pointer instead of intializing
it in the pointer afterwards. (This is dumb and inferior, though
possibly will be optimized away, but Coverity complains about
initializing *seq_new with an uinitialized member that is immediately
overwritten otherwise.) Reorder the struct members to put the nstime_t
first and group the 32 and 16 bit items together, which is marginally
better on platforms with 8 bit alignment.

Coverity CID 1499435