VoIP Calls: Check whether Q.931 calls are complete
Process the queue of VoIP calls backwards from the tail, since the most
recent calls are added to the end. (This should probably be done
elsewhere in the function.) If the most recent matching call (same CRV,
etc.) is active, or the new message type is for a call ending, then
use that call. Else, the newly tapped message creates a new call.

Fix #19653