asn2wrs: Don't warn about identical duplicate directives in quiet mode
If a directive is a duplicate of another, but the parameters (outside of
the location in the set of conformation files) are the same, then it's
pedantic to warn about them and we don't need to do so when in quiet
mode. A common case is when an ASN.1 module's protocol name is given
multiple times, with #.MODULE in its own conformation file, but also
with the #.MODULE_IMPORT directive in conformance files of other modules
that import it.

We *could* remove the redundant declarations eventually, but we don't
want to clutter compared to more significant warnings (which are still
sometimes intentional.)