epan: Use unsigned offsets with tvb_ensure_bytes_exist
There are no cases in the repository of this function being used
with a negative offset. A negative offset doesn't make sense to
use, especially as a negative offset is from the end of the captured
data (not reported data, as one might think).  There's always a
more sensible way to express that.

The one function that calls it internally, tvb_get_raw_string,
already throws a DissectorError if its offset parameter is < 0.

Ping #20103