epan: Use unsigned offsets with tvb_bytes_exist and _offset_exists
There are no cases in the repository where this is used with a negative
offset. Because a negative offset means "from the end of the captured
data" (not the reported data, as one might think), using it with a
negative offset probably doesn't do what someone expects, and there's
usually a more sensible way to express the same idea.

Ping #20103