epan: Use unsigned offsets with tvb_*_length_remaining
There has never been much reason to use negative offsets with
tvb_captured_length_remaining, and there do not appear to be any
such uses in the repository. It's possible to perhaps construct
a reason with tvb_reported_length_remaining, but only if the user
knows that negative offsets refer to the captured length, and it's
still not the simplest way to do it. There don't appear to be any
such uses either.

It is much more likely that a negative offset passed in is the
result of overflow (positive overflow, or negative after subtraction)
and treating as a value from the end of the captured length can lead
at worst to infinite loops.

Ping #20103