Fix some incorrect uses of -1 as offset
In gpef, this is apparently meant to be an item covering the entire tvb.

In RTP ED137, these are apparently meant to create a generated item one
byte back out of the subset tvb into the parent tvbuff. It doesn't work
that way, so just make it a zero offset, zero length generated item for
now.

Ping #20103