infiniband/smbdirect: fix dissecting the first smbdirect negotiate request
Here's a sample capture
https://gitlab.com/wireshark/wireshark-web/-/issues/62
https://gitlab.com/-/project/22852016/uploads/538c095d29f38f95b5231fd7b9e05bdf/rdma2-rxe-r6.18-tmp-10.pcap.gz
This is a Windows server 2025 as server with an Intel-E810-CQDA2 card in
roce v2 mode (10.123.102.5) and smbclient kernel smbdirect using
software roce rdma_rxe (10.123.102.4) This uses smb signing so there's
no rdma offload for the write.

This fixes dissecting of frame 6, while reassembling still works
at the infiniband layer, e.g. frames 50 + 51, as well
as the smbdirect layer, see how frame 51 is part of frame 2388.

Signed-off-by: Stefan Metzmacher <metze@samba.org>