epan: Use unsigned offset and length for tvb_memcpy, tvb_memdup
There are no uses in the repository or unsigned offset or length
for these functions. Since anything retrieving all of the captured
data would still need to get the length from tvb_captured_data_remaining
(since this isn't a null-terminated string or otherwise has a known
length), it's not clear if we need a _remaining function to replace
theoretical uses of -1 to mean "to the end of the captured data."

However, note that there are uses of -1 as a length with tvb_get_ptr,
which is similar and where the same concerns would apply. All of those
cases could be rewritten easily, though.

Ping #20103