Qt: Don't pass -1 as a length to tvb_get_ptr to copy a single field
When Showing or Exporting Packet Bytes, we get the raw packet bytes
of a single field. If that field is part of a composite tvb, then
passing -1 to tvb_get_ptr does an expensive memcpy of all the bytes
instead of just the necessary bytes (or no copy at all, if the needed
bytes are part of a single real component tvbuffer.)

It's less important for a decompressed tvb, but go ahead and eliminate
the -1 anyway to reduce usage of that idiom. Change the length there to
captured length, which is correct (though also doesn't matter because
no decompression function should return a tvb with captured length
less than reported length.)

We still have to explicitly pass the length to various other functions,
so if the tvb_get_ptr call fails, then we would be committing a buffer
overflow anyway.