epan: Use ws_abort_if_fail to satisfy static analysis in mark_truncated
There are problems with mark_truncated and the string is not necessarily
properly null truncated if the size passed in is too small. That should
never happen, but Coverity and other static analyzers don't believe that
it halts when ws_assert is used. Use an unconditional assert to try to
hint to the static analysis.