netscaler: Fix absolute time when saving v2.0-v2.2, at least when not seeking
For the Netscaler v2.0 file type, correctly write the absolute time to the
32-bit absolute time field, instead of to the relative time field. Also,
the size of the absolute time in v2.0-v2.2 Netscaler files is 32-bit, so
fix that (which affects the memcpy when running on Big Endian
architecture.)

On versions 2.3 and later, this was written incorrectly as well but a more
precise 8 octet absolute time is used instead when reading.

Coverity CID 1499452

Fixup c317a1c3cdee6c173038dc28e60c2199803dd6a3

This fixes saving from editcap, one-pass TShark, etc. For Export
Specified Packets from Wireshark and two-pass TShark, this makes the
absolute timestamps 0 instead of random, due to a different issue where
for Netscaler before v2.3 (including v1) timestamps in the packet are
relative and only available in the wtap_rec from libwiretap on a
sequential read, not when using wtap_seek_read.

That could possibly be fixed by adding the timestamp back from the
frame data, where it is saved for this reason and made available
to the packet_info struct (cf. epan/packet.c dissect_record()).
Add some comments in the relevant places in tshark.c and file.c
about that.