wsutil+macOS: Add is_plugin_filename and disable some plugins
Add is_plugin_filename, which returns true if a filename ends with .dll
on Windows or .so on other platforms. Use it to check for valid filenames
when loading Falco plugins.

In GitLab CI, disable a the Falco gcpaudit and k8saudit plugins by
appending ".disabled" to their respective filenames.

Note in the release notes that this works around #20869