SMB SMB2: New and fixed access masks
FIXED: File access mask bit definitions in Create AndX requests
and responses were being used for directories as well as for
files/named pipes. This was also the case in SMB2 Tree Connect
requests. The directory access bit table was added per MS
specifications. Open AndX responses call dissect_smb_fid()
which has been changed to accomodate them. Fixed the meaning
of the "Non-Directory" Create Option.

NEW: In SMB, a distinction is made between a clients desired
access rights and the server's maximal granted access rights
and are labeled accordingly. The desired and maximal are
ANDed to produce the granted access rights. The desired access
and granted masks were added to the GUID subtree in SMB2 for
opens, Reads and Writes.

NOTE: SMB2 uses the the same routine as CIFS/SMB to dissect
access masks so these changes are reflect in SMB2 as well.

Fixed more pipeline errors-4