Various dissectors: FT_STRINGZ fixes
Fix places where dissectors call tvb_get_stringz_enc only for the
length, in some cases where they call that for the length and add
a field using the length, effectively retrieving the string twice.
There's also LBMSRS, which uses a belt-and-suspenders approach of
having the string length encoded but also including a trailing
null byte (the length includes the nul.) In that one the string
was also being retrieved twice.

(Fun fact: I grepped for "belt-and-suspenders" and the only other
protocol with that phrase, packet-ajp13.c, uses it in the same
context of a string with both a length prepended and a trailing
null, with the null included in the length.)

Use proto_tree_add_item_ret_string, proto_tree_add_item_ret_length,
or proto_tree_add_item_ret_string_and_length to eliminate the
double string processing.

In some cases, convert fields from FT_STRING to FT_STRINGZ. In one
case, convert to FT_STRINGZTRUNC (as it appears to have a maximum
size.)

Also fix a few other tree items in the Gryphon plugin.

Ping #20103 (this is preliminary for converting tvb_get_stringz_enc).