HTTP URL Encoded: Keys and values are not FT_STRINGZ
The keys and values in HTTP form URL encoded are not null terminated
in the packet, so they are not FT_STRINGZ.

This doesn't in practice affect anything right now, but it might make
sense in the future to use a length of 0 for FT_STRINGZ to mean
"unknown, search for the terminator" in the same way that -1 does now,
in order to allow switching to unsigned lengths. A real FT_STRINGZ
should never have length zero. If a protocol has "a null terminated
string, but also it can be completely missing" there should be some
other way to determine that it is missing, as searching for the NUL
will cause an exception. That other method likely means that the field
should really be a FT_STRINGZTRUNC or FT_STRINGZPAD instead.