SABP: Get rid of some code that doesn't do anything
New-Serial-Number is a Serial-Number, so the way asn2wrs generates the
code the VAL_PTR for New-Serial-Number is unusued. Everything is handled
as desired (with the proper hf) in dissect_sabp_Serial_Number.

The only slight difference is that the separate ett isn't used, but it's
never been used.

Coverity CID 281372