epan: Implement tvb_find_uint8_remaining returning a bool
Since tvbuff_t internally always have unsigned offsets and lengths,
have the internal tvb_find_uint8 operator set the found offset in
a pointer to unsigned and return whether or not it was found as a bool,
instead of returning an int where -1 means not found.

Expose publicly new API functions, one that searches to the end of
the captured bytes in the tvbuffer and one that searches up to a maxlength
or the captured bytes, whichever is shorter, both of which return a bool
and place the found offset into a pointer to an unsigned. In the case
where it is not found, the bool returned is false but the pointer to the
offset is set to one past the last offset checked. This should decrease
the chance of bad behavior caused by not checking the return value for
-1 with the existing function.

Use the new function internally in a couple of places.

Ping #20103