linx: Use add_item_ret_length instead of an extra tvb_strnlen
If adding a FT_STRINGZ, it is more efficient to use the _ret_length
function than to retrieve the length a second time when the tree is
visible. Also, the existing code was incorrect because it failed to
advance past the terminating NULL, as tvb_strnlen (but not tvb_strsize)
does not include the '\0' terminator, though in the sample capture
files those strings were at the end.