QT: Document SCTP Graph Dialog using 2 FAMs in one struct in C++
Flexible Array Members are not legal C++, and I'm surprised that the
compilers don't complain about this since it's using the "length 1"
approach instead of [0] or []. They might complain at runtime after
recent warning changes, if I can find any samples of SCTP with gap
reports and duplicates TSNs in the SACK headers.

Change something where instead of taking the address of a singleton
and doing some pointer arithmetic because it happened to have the
right type as the unnamed FAM off the end that we wanted to address
(Coverity CID 1439124 did not like this), instead find the address
of the end of the first FAM and then cast it to a different type
of array. (What will Coverity think of this?)

The code is still unsafe and I'm not sure it works on fuzzed data
or even some truncated data.