Diameter: Flag and return on bad too short lengths
RFC 6733 AVP Header:

   AVP Length

      The AVP Length field is three octets, and indicates the number of
      octets in this AVP including the AVP Code field, AVP Length field,
      AVP Flags field, Vendor-ID field (if present), and the AVP Data
      field.  If a message is received with an invalid attribute length,
      the message MUST be rejected.

If the AVP length is less than that, add an expert info and reject. On
Wireshark 4.6 and earlier, this causes tvb_new_subset_length to throw
an exception with a negative length, but with length unsigned that does
not happen on the current development branch. (The expert info is
better anyway.)

Fix #20914