btatt: Enable the detection of ATT transaction timeouts
This basically fixes / enhances the transaction tracking so that
requests/indications point to responses/confirmations and vice-versa.

An expert info item is attached to a packet if no response/confirmation
has been observed for that packet so far, so that users may filter for
`btatt.incomplete_transaction` to identify possible ATT timeouts.