PCNFSD: Improve deobfuscation of identity and password
In one case, a trailing null was not being appended but then strlen
was used on the result (Coverity CID 1676267.) There were also
issues around deobfuscating UTF-8 REPLACEMENT CHARACTERS in the case
of fuzzed data.

In another case, the function use was not updated for how the
deobfuscation no longer works in place after
commit 8ed7a73e22c049a2e013bb436e599bff41fc5b9b, with the result
that the deobsfucation was not actually doing anything for the password.