req_resp_hdrs: Pass the desegment condition into tvb_find_line_end
We can only desegment if the buffer is not truncated. Pass that
condition into tvb_find_line_end instead of testing it afterwards.
In the case where the buffer is truncated, this causes tvb_find_line_end
to return the captured length remaining instead of -1, which is better
for passing to tvb_get_string_enc. It has the same result as when -1
for the end of the [captured] buffer was allowed.

Coverity CID 1676195

Also, change two places where packet-http and packet-rtsp programmatically
left the reported_length as -1 when not found in a packet, and depended
on tvb_new_subset_length treating -1 as meaning "to the end of the
packet."