epan: Don't overflow LZNT1 decompression at output offset 0
From MS-XCA 2.5.1.4:

Let U be the amount of uncompressed data that has already been processed in the
current chunk… Then let M be the largest value in [4…12] such that 2M-1 < U,
or 4 if there is no such value.  A compressed word then has the format D = M
and L = 16 – M, with the displacement occupying D high-order bits and the length
occupying L low-order bits.

The current implementation takes an approach of repeatedly dividing U by
2 in order to find M, which requires (due to truncation with integer
division) subtracting 1 from U and using a weak inequality instead of a
strict one. However, this has overflow at the base condition of U == 0.

Coverity CID 1604219

Also, change a few variables which must be unsigned to unsigned types.