S1AP: Fix error handling and setting offset with TRY..CATCH
We want to set the offset at the end to the dissected offset.
Change SourceeNB-ToTargeteNB-TransparentContainer to match
TargeteNB-ToSourceeNB-TransparentContainer. Note that the
end offset was not actually used here, because it was assigned
to the non-volatile offset value which was then overwritten
back to its starting value.

Coverity 1517142

Fixup 16ea22eaa4dcf579da33b8767a2958fe9e00626a