MEGACO: Pass a tvb subset into dissect_megaco_descriptors
This will naturally terminate the various find commands, so we won't
need some checks for if we found something outside the descriptors
end offset. This doesn't use the new tvb_find_ functions yet, so it
could be backported. We might want to use those in the future to
simplify the checks for offset -1.

Coverity CIDs 1604264 and 1676261