dfilter: Convert FT_SCALAR to FT_UINT64 for non Number types
Character constants, and literals and strings (the latter two if
there's a value string) can be converted to FT_INT64 but not
FT_DOUBLE. So if a FT_SCALAR is requested, try to convert to
FT_INT64. This prevents some crashes with filters that try to
divide a FT_ABSOLUTE_TIME or FT_RELATIVE_TIME by character constants,
literals, or strings:

frame.time_relative > ${frame.time_relative} / 0.:5
frame.time_relative > ${frame.time_relative} / "foo"
frame.time_relative > ${frame.time_relative} / '\x47'

Before:
        [(none) ERROR] epan/ftypes/ftypes.c:503 -- fvalue_new(): assertion failed: ftype < FT_NUM_TYPES

[Compiled in Debug mode; a different error if compiled in Release mode
where asserts are not compiled in]

for all three.

After:
$ ./run/dftest $(cat bad-filter-crash2)
Filter:
 frame.time_relative > ${frame.time_relative} / "foo"

Error: Signed integer (64 bits) cannot be converted from a string ("foo").
  frame.time_relative > ${frame.time_relative} / "foo"
                                                 ^~~~~

for the first two and success for the third:

./run/dftest $(cat bad-filter-crash3)
Filter:
 frame.time_relative > ${frame.time_relative} / '\x47'

Instructions:
 0000 READ_TREE        frame.time_relative -> R0
 0001 IF_FALSE_GOTO    6
 0002 READ_REFERENCE   ${frame.time_relative} -> R1
 0003 IF_FALSE_GOTO    6
 0004 DIVIDE           R1 / 71          -> R2
 0005 ANY_GT           R0 > R2
 0006 RETURN

The error messages are perhaps not ideal, but this is better than
crashing.