wps: Simplify logic to prevent future null pointer dereference
If the default actually is reached, and fmt stays NULL on a field
that has a strings entry, that leads to a null pointer dereference.
That doesn't happen with any existing fields that are registered,
but it's better to clean up the logic. We only want to append anything
at all if fmt is not NULL, and that will only happen if hf_info is
not NULL, so move around the tests.

Coverity 1604183