UET: Retrieve the flags and use them
In the standard request, flags was being set to 0 and never retrieved,
but used in a test.
In the medium request, flags (.hd) was being set to false and never
retrieved.

Use proto_tree_add_bitmask_ret_uint64 to retrieve the flags in both
cases and use them properly. Also use the #defines for the bitmasks
when defining the fields.

Coverity CIDs 1676242 & 1676238