idl2wrs: Fix a check for a single bit set in a bitmap
The check (val & (val-1)) is intended to be nonzero for all cases
other than where a single bit is set, but it doesn't correctly handle
the case where val is zero, which also ends up being zero. Zero isn't
a valid value for a bitmap value; one bit needs to be set. (Not that
0 is used in the idl processed by this.)

Checking for it also avoids a warning about overflow from Coverity.
Coverity CID 1604196