p4rpc: Fixup truncation
Don't call ws_utf8_truncate unless we wrote to the end of the buffer
(and thus there's possibly a truncated UTF-8 codepoint); this prevents
UB with the end of the string buffer not being initialized.

This shows a reason why snprintf is tricky, remembering which parameters
include the null terminator and which ones don't.

Fix #20922 (while keeping #20916 fixed.)