Have dissectors reduce rather than expand the protocol item length
Analysis engines like MATE depend on dissectors which are uncertain
of their PDU length up front to first declare the protocol item to
span the entire TVB, then at a later stage reduce their TVB scope.

With this change a few remaining dissectors not following this pattern
are adjusted to work like this.