TRDP: switch the dissector to the TCNOPEN-community origined version
The current dissector is replaced
by the so far out-of-tree-plugin "TRDP SPY". Most notable change is the
support for payload definitions based on XML-files as defined in the protocol's
standard IEC 61375-2-3. The dissector includes a base set of standardized
data-sets (in the resource folder). It can be extended by loading custom XML
files through the dissector properties. There are further tweaks possible
through the settings. Currently, the dissector does not support conversations
and neither deep-dives into illegal property values or safety related protocol
faults.