DOF: Avoid tvb_get_ptr with length -1
Call tvb_get_ptr with the same length that is used later on in decrypt()
as the length of the buffer. This prevents copying unnecessarily (e.g.,
in a composite TVB), and also guarantees that there's no memory error
later by accessing past that length by throwing an exception. An
exception was probably thrown by this point if offset was past the
captured data, but ensure it.