epan: Return unsigned from tvb_raw_offset
This requires sorting tvb->raw_offset as unsigned. Instead of using
-1 to mean "hasn't been calculated yet", add a flag for if it has
been calculated.

When creating a subset tvb, if the raw offset has been calculated
for the parent, go ahead and calculate it for the subset. (We'll
probably need it later, and this saves some recursion later.)
We probably actually can calculate the raw offset for all the types
of tvbuffs we have upon creation (a composite is 0), which might
make sense.

Add some notes to the documentation about the small differences
between tvb_offset_from_real_beginning and tvb_raw_offset.