capture: Use pcap_init on Windows if we have it
On Windows, use pcap_init if present to request that error strings be
returned in UTF-8 instead of the local code page, so we don't have to
translate them. This also keeps pcap_create from trying to determine
if the device string passed in is in UTF-16LE (which the deprecated
and disabled by this option pcap_lookupdev would return).

This function was first added in libpcap 1.10.0